Re-Typed STS RunInfo "Run Number" Usage... 	- was erroneously typed as "unsigned long", really is "uint32_t"... 		-> fixed boost::lexical_cast<> from RunInfo XML parsing... 		-> fixed initial STS Log Message with Run Number... ("%u") 		-> changed type of STS RunInfo::getRunNumber() method...

+++ sts/StreamParser.cpp
@@ -1359,7 +1359,7 @@ StreamParser::rxPacket
                         (const xmlChar*)"run_number" ) == 0 )
                 {
                     m_run_info.run_number =
-                        boost::lexical_cast<unsigned long>( value );
+                        boost::lexical_cast<uint32_t>( value );
                 }
                 else if ( xmlStrcmp( node->name,
                         (const xmlChar*)"proposal_id" ) == 0 )
@@ -3217,7 +3217,7 @@ StreamParser::receivedInfo( InfoBit a_bit )
         m_info_rcvd |= INFO_SENT;
 
         syslog( LOG_INFO,
-            "[%i] target_station: %u, beam: %s:%s, prop: %s, run: %lu",
+            "[%i] target_station: %u, beam: %s:%s, prop: %s, run: %u",
             g_pid,
             m_run_info.target_station_number,
             m_run_info.facility_name.c_str(),
